Moved some classes out of the PowerFilter package and into Core proper.
